Add oop-demo-shapes module with Shape, Circle, Square and tests
