+++ oop-demo-shapes/src/main/java/com/springliviu/shapes/Circle.java
@@ -1,4 +1,14 @@
 package com.springliviu.shapes;
 
-public class Circle {
+/** Круг */
+public class Circle implements Shape {
+    private final double radius;
+    public Circle(double radius) { this.radius = radius; }
+
+    @Override
+    public String name() { return "Circle"; }
+
+    @Override
+    public double area() { return Math.PI * radius * radius; }
 }
+
